refactor(org): add org requests to dedicated file
