+++ src/api/user.ts
@@ -1,5 +1,6 @@
 import { formatQueryParameters, isValidId, parseSnakeCaseToCamelCase } from "../api"
 import {
+    BadRequestException,
     CreateUserException,
     UpdateUserEmailException,
     UpdateUserMetadataException,
@@ -461,4 +462,38 @@ export function disableUserCanCreateOrgs(authUrl: URL, integrationApiKey: string
 
         return true
     })
+}
+
+export type InviteUserToOrgRequest = {
+    orgId: string
+    email: string
+    role: string
+}
+
+export function inviteUserToOrg(
+    authUrl: URL,
+    integrationApiKey: string,
+    inviteUserToOrgRequest: InviteUserToOrgRequest
+): Promise<boolean> {
+    const body = {
+        org_id: inviteUserToOrgRequest.orgId,
+        email: inviteUserToOrgRequest.email,
+        role: inviteUserToOrgRequest.role,
+    }
+
+    return httpRequest(authUrl, integrationApiKey, `/api/backend/v1/invite_user`, "POST", JSON.stringify(body)).then(
+        (httpResponse) => {
+            if (httpResponse.statusCode === 401) {
+                throw new Error("integrationApiKey is incorrect")
+            } else if (httpResponse.statusCode === 400) {
+                throw new BadRequestException(httpResponse.response)
+            } else if (httpResponse.statusCode === 404) {
+                return false
+            } else if (httpResponse.statusCode && httpResponse.statusCode >= 400) {
+                throw new Error("Unknown error when inviting a user to the org")
+            }
+
+            return true
+        }
+    )
 }
